backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.0",
  "python-dotenv",
  "fastmcp>=2.0",
  "anthropic>=0.40",
  "agentmail",
  "modal",
  "hubspot-api-client>=12.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]
